requirements.txt
streamlit==0.71.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
Pillow==8.0.1
keras==2.4.3
numpy==1.18.5
opencv-python==4.1.0.25
matplotlib==3.1.0
